edit trip with images changes

+++ client/src/pages/MyTrip.jsx
@@ -7,12 +7,12 @@ function MyTrip() {
     const navigate = useNavigate();
     const [trip, setTrip] = useState([]);
     const [isEditFormOpen,setIsEditFormOpen]= useState(false);
-    const [image, setImage]= useState(null);
+    // const [image, setImage]= useState(null);
     
     //actualiza la constante myTrips and Images
     useEffect(() => {
         getTrip();
-        getImage();
+        // getImage();
     }, []);
 
     //llama a la base de datos y trae todos los viajes
@@ -20,30 +20,32 @@ function MyTrip() {
         fetch(`http://localhost:5000/api/trips/${trip_id}`)
             .then((response) => response.json())
             .then((data) => {
-                setTrip(data[0]);
-                setEditingTrip(data[0]); // Asigna los detalles del viaje a editing Trip
+                setTrip(data);
+                // setEditingTrip(data); // Asigna los detalles del viaje a editing Trip
             })
             .catch((error) => {
                 console.log("No trips available")
             });
     };
 
-    //llama a la base de datos y trae el iamgen del trip
-    const getImage = () =>{
-        fetch(`http://localhost:5000/api/trips/images/${trip_id}`)
-        .then((response) => response.json())
-        .then((data)=>{
-            setImage(data[0]);
-        })
-        .catch((error) => {
-            console.log("Oops! Something went wrong")
-        });
-    }
+    // //llama a la base de datos y trae el iamgen del trip
+    // const getImage = () =>{
+    //     fetch(`http://localhost:5000/api/trips/images/${trip_id}`)
+    //     .then((response) => response.json())
+    //     .then((data)=>{
+    //         setImage(data[0]);
+    //     })
+    //     .catch((error) => {
+    //         console.log("Oops! Something went wrong")
+    //     });
+    // }
 
     const handleDelete = async (trip_id) => {
         try {
+            if(trip_id){
             const response = await fetch(`http://localhost:5000/api/trips/${trip_id}`,
                 { method: 'DELETE' });
+            }
             if (response.ok) {
                 navigate(`/trips`);
                 return;
@@ -54,7 +56,7 @@ function MyTrip() {
         };
     };
 
-    const handleEdit = (id) => {
+    const handleEdit = () => {
         setIsEditFormOpen(true);
         // Navega a la página MyTripAdd y pasa el trip_id como parámetro para la edición
         navigate(`/mytripadd/${trip_id}`);
@@ -65,7 +67,7 @@ function MyTrip() {
                 <div key={trip.id}>
                 <h1 className="pt-14 pb-20">{trip.name}</h1>
                 <div className="">
-                {image && <img src = {`http://localhost:5000/images/${image.name}`} id={image.id} alt={image.description}/>}
+                {trip && <img src = {`http://localhost:5000/images/${trip.imageName}`} id={trip.imageName} alt={trip.imageDescription}/>}
                 </div>
                     <p className="flex mb-8"></p>
                     <p>{trip.description}</p>
